requirements.txt
numpy>=1.26.0
pandas>=2.2.0
python-dateutil>=2.9.0
PyYAML>=6.0.0

# Ingestion schema (kept for ADLS/local ingestion CLI)
pydantic>=2.6.0

# ADLS uploads (used by --to-adls)
azure-storage-blob>=12.21.0
azure-identity>=1.16.0

# Modeling and serving
scikit-learn>=1.3.0
xgboost>=2.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
requests>=2.31.0
